Create Transfer_Xception.ipynb
Transfer Learning with keras pretrained model- Xception
This model has weights that have been pretrained on imagenet
More layers are added on top of the model.
Could use data augmentation for better results- I did not try that yet
This still need to be fine-tuned for better performance

There are other pretrained models we can try if this doesnt seem like it is working

Some other things that could be done is geting all the mixed dog breed images to arrays and creating checkpoints. The best weight file is need for the predictions.