Update Kaggle Preprocessing, Added train.zip
Added header and comments to kaggle preprocessing. Added a commented out method to extract images from a zipped version of the train file (I am unsure if Alan can run this without this. I can't, but I might be missing something). If we do not want these, comment can be deleted later.